feat: Move out MDsveX options to a separate file
